+++ controllers/delivery-taxes-controller.js
@@ -22,11 +22,14 @@ const getAllDeliveryTaxes = async (req, res, next) => {
 };
 const getSingleDeliveryTaxes = async (req, res, next) => {
   const { id: deliveryTaxesId } = req.params;
+  if (!deliveryTaxesId) {
+    throw new BadRequestError("Please enter a delivery taxes id");
+  }
   const deliveryTaxes = await prisma.deliveryTaxes.findUnique({
     where: { id: deliveryTaxesId },
   });
   if (!deliveryTaxes) {
-    throw new BadRequestError("Delivery Taxes not found");
+    throw new NotFoundError("Delivery Taxes not found");
   }
   return res
     .status(StatusCodes.OK)
@@ -52,18 +55,6 @@ const createDeliveryTaxes = async (req, res, next) => {
   if (!zodModel.success) {
     throw new BadRequestError(zodModel.error.errors[0].message);
   }
-  // const originState = await prisma.state.findUnique({
-  //   where: { id: originStateId },
-  // });
-  // if (!originState) {
-  //   throw new BadRequestError("Origin State not found");
-  // }
-  // const destinationState = await prisma.state.findUnique({
-  //   where: { id: destinationStateId },
-  // });
-  // if (!destinationState) {
-  //   throw new BadRequestError("destination State not found");
-  // }
 
   const createdDeliveryTaxes = await prisma.deliveryTaxes.create({
     data: {
@@ -91,10 +82,8 @@ const updateDeliveryTaxes = async (req, res, next) => {
     additionalFeesAfterKg,
     feePerKg,
   } = req.body;
-  if (!id) {
-    throw new BadRequestError("Please send a delivery taxes ID");
-  }
   const zodModel = UpdateDeliveryTaxesZodModel.safeParse({
+    id: id,
     originStateId: originStateId,
     destinationStateId: destinationStateId,
     baseFee: baseFee,
@@ -104,22 +93,6 @@ const updateDeliveryTaxes = async (req, res, next) => {
   if (!zodModel.success) {
     throw new BadRequestError(zodModel.error.errors[0].message);
   }
-  // if (originStateId) {
-  //   const originState = await prisma.state.findUnique({
-  //     where: { id: originStateId },
-  //   });
-  //   if (!originState) {
-  //     throw new BadRequestError("Origin State not found");
-  //   }
-  // }
-  // if (destinationStateId) {
-  //   const destinationState = await prisma.state.findUnique({
-  //     where: { id: destinationStateId },
-  //   });
-  //   if (!destinationState) {
-  //     throw new BadRequestError("destination State not found");
-  //   }
-  // }
 
   const deliveryTaxes = await prisma.deliveryTaxes.update({
     where: { id: id },
@@ -138,7 +111,7 @@ const updateDeliveryTaxes = async (req, res, next) => {
     },
   });
   if (!deliveryTaxes) {
-    throw new BadRequestError("Delivery Taxes not found");
+    throw new NotFoundError("Delivery Taxes not found");
   }
 
   return res.status(StatusCodes.OK).json({
@@ -150,11 +123,14 @@ const updateDeliveryTaxes = async (req, res, next) => {
 
 const deleteDeliveryTaxes = async (req, res, next) => {
   const { id: deliveryTaxesId } = req.params;
+  if (!deliveryTaxesId) {
+    throw new BadRequestError("Please enter a delivery taxes id");
+  }
   const deliveryTaxes = await prisma.deliveryTaxes.delete({
     where: { id: deliveryTaxesId },
   });
   if (!deliveryTaxes) {
-    throw new BadRequestError("Delivery Taxes not found");
+    throw new NotFoundError("Delivery Taxes not found");
   }
 
   return res
